Refactor connectors into a single package
